Python/get_brokers_list/requirements.txt
# Function dependencies, for example:
# package>=version
#Flask==0.10.1
Flask==2.0.1
#google-cloud-storage==1.28.1
google-cloud-bigquery==1.24.0
pandas==0.24.0
numpy==1.19.0